requirements/local.txt
-r base.txt

django-coverage-plugin
django-debug-toolbar
flake8
pytest==4.6.4
pytest-cov
pytest-django
pytest-factoryboy
pytest-mock
pytest-sugar
pytest-xdist
requests-mock
